Add very simple tab-completion to the toplevel
This is based on uTop's implementation, with all of the utf-8 parts
stripped out. Unfortunately compiling camomile with js_of_ocaml was
causing an error early during initialisation.

I've kept the uTop modules in the same structure they were originally
so it's easy to reinstate if camomile gets fixed or if we find a
suitable replacement.

The UI/UX/DX could do with some love!