use 4.14.0 as latest ocaml version (#470)
* use 4.14.0 as latest ocaml version

To get up and running it's confusion if an older version is shown. I've also added an instruction to show all possible switches but as that returns a huge list, I'm wondering if that might be too much for the up and running. On the other hand, it would be good for new people to find a list of possible releases they can use.

see for example this question: https://discuss.ocaml.org/t/are-ocaml-install-instructions-up-to-date/9941