Generate canonical links
Pages accessible using different urls shall include a canonical
link, indicating which url is the best (search Google for “consolidate
duplicates urls”). This is useful for web crawlers. Ahrefs reports
“Duplicate pages without canonical” as an issue. This patch adds
canonical links to several pages.

Main tag generation takes place in layout.eml.

Often, it is sufficient to pass the right url to Layout.render. This is
the case for:

* about.eml
* academic_users.eml
* best_pratices.eml
* blog.eml
* books.eml
* community.eml
* industrial_users.eml
* jobs.eml
* learn.eml
* news.eml
* packages.eml
* papers.eml
* platform.eml
* problems.eml
* releases.eml
* tutorial.eml

A few others are almost the same, except some obvious path part needs
to be appended. This is the case for:

* news_post.eml
* release.eml
* success_story.eml
* workshop.eml

Other changes includes parameter forwarding:

* learn_layout.eml
* ocamlorg_frontend.ml
* page.eml
* tutorial.eml

In package_search.eml, the same canonical is used as in packages.eml.

In home.eml, the empty string is used to have "https://ocaml.org" be
the best url (without trailing slash)

In package_version.eml, package version is not used in order to
have the top level pages being referenced.

In handler.ml a combination of parameter forwarding and canonical
url setting takes place.

In playground.eml, the string "https://ocaml.org" is hardwired a
second time because it uses an ad-hoc layout.

No canonical link is generated in not_found.eml, on purpose.