Generate canonical links (#587)
Pages accessible using different URLs shall include a canonical
link, indicating which URL is the best (search Google for “consolidate
duplicates URLs”). This is useful for web crawlers. Ahrefs reports
“Duplicate pages without canonical” as an issue. This patch adds
canonical links to several pages.

Main tag generation takes place in layout.eml.

Often, it is sufficient to pass the right URL to `Layout.render`. This is
the case for:

* about.eml
* academic_users.eml
* best_pratices.eml
* blog.eml
* books.eml
* community.eml
* industrial_users.eml
* jobs.eml
* learn.eml
* news.eml
* packages.eml
* papers.eml
* platform.eml
* problems.eml
* releases.eml
* tutorial.eml

A few others are almost the same, except some obvious path part needs
to be appended. This is the case for:

* news_post.eml
* release.eml
* success_story.eml
* workshop.eml

Other changes include parameter forwarding:

* learn_layout.eml
* ocamlorg_frontend.ml
* page.eml
* tutorial.eml

In home.eml, the empty string is used to have "https://ocaml.org" be
the best URL (without trailing slash)

In handler.ml a combination of parameter forwarding and canonical
URL setting takes place.

In playground.eml, the string "https://ocaml.org" is hardwired a
second time because it uses an ad-hoc layout.

No canonical link is generated in not_found.eml, on purpose.