Update Opam repo to update Hilite to 0.2.0
In order to update hilite from 0.1.0 to 0.2.0 the hash used to
point to a sufficient and known-to-work commit in opam-repository
had to be refreshed. Merge of hilite-v0.2.0 was used.

However, lifting opam repo hash caused a docker image compilation
failure at stage 6: opam install . --deps-only
This was worked around by raising alpine from 3.14 to 3.17
at stage 1.

However, this caused a crash when running the built container.
This was worked around by raising alpine from 3.12 to 3.17
at stage 8.

Further investigation are required in order to understand the
failures, which are likely to kick-in again.