Add XenServer (Cloud Software Group) job
XenServer (formerly part of Citrix) has an open position for work on its
Toolstack, which is written in OCaml.
Citrix is now part of Cloud Software Group,
but XenServer is no longer part of Citrix: it is part of Cloud Software
Group directly, so use Cloud Software Group as the company name.

For now this links to the logo on cloud.com while we are still sorting
out a new XenServer website.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>