Scraper: Update watch_scrape.ml after addition of yoshi for parsing YAML
When changing the YAML parsing in watch.ml to use the yoshi tool,
watch.yaml was turned into a list of watch entries, instead of
this list being part of a struct.

watch_scrape.ml was not updated accordingly. This patch updates
watch_scrape.ml to emit a YAML list of watch entries.