Introduce the OCaml Changelog (#1109)
This commit introduces the OCaml Changelog - a feed of the newest releases and
feature highlights of the OCaml Compiler, OCaml Platform and other official
projects including Ppxlib and Omd.

It is served at https://ocaml.org/changelog.

The goal of the Changelog is to offer a page where OCaml developers can stay up
to date with the latest official releases, and most importantly quickly learn
about the most important new features available and how to use them.

At this stage, the OCaml Changelog is an experiment. We're merging as such to
facilitate communication with everyone involved, especially the maintainers of
the projects featured in the Changelog.

In particular, if we decide on supporting the OCaml Changelog in the long term,
we'll need to find a good release process with the maintainers of the tools that
are featured in the Changelog, to make sure that every release gets an
announcement and that important features are highlighted.

We have talked about this with several maintainers, but at this point, we have
not agreed on a definitive process. While the Changelog remains in an
experimental stage, for demonstration purposes, we intend to take on the
workload of creating the changelog entries and working with the maintainers to
write the announcements, until we've built a consensus on the final process.

Note that the introduction of the Changelog impacts the content that is
available on the OCaml Blog.

The previous News section, which contained the newsletters (Multicore and
Compiler), but also releases and articles from the OCaml Platform Blog, now only
contains newsletters. The Platform Blog articles are now part of the OCaml
Planet feed (the aggregate of community blog posts), and the releases are part
of the OCaml Changelog.

The commit also upgrades the opam-repository to
9cf27a6683dc4e56ccf7e35c008198828bd5c267

---------

Co-authored-by: Sabine Schmaltz <sabine@tarides.com>