Document the OCaml Platform governance
This commit extends the OCaml.org governance to include the OCaml Platform. We
document the lifecycle and requirements for the OCaml Platform tools.

We also re-organise the structure of the governance to move the
OCaml.org-specific text in a single section.

Outside of minor text changes to mention the OCaml Platform, the following
sections remain untouched:

- Introduction
- Roles and responsibilities
- Processes

We also rename the document from "OCaml.org Governance" to "OCaml Governance" to
indicate that it is no longer specific to the OCaml.org project and sub-domains.

Following the goal of the document to describe the _existing_ governance and not
what we wish it was, the lifecycle and requirements for the Platform tools are
taken from how the OCaml Platform is currently governed. Both the lifecycle and
requirements have also been laid down in
[this talk](https://watch.ocaml.org/w/2KbfRNv2oLtkKXkbd5u9F1).