Docker build and push image container registry
Yaml file was generated by  ChatGPT using the following prompt:

Write a yaml file which implements a github workflow action. The action is only triggered manually. The user must select a branch, default is main. It triggers build of the project Docker file. The built image is then push to github container registry using a GITHUB_TOKEN.

As of this commit, it hasn't been touched