How to Write an OCaml Program
This tutorial is a follow-up from “A Tour of OCaml”, it contains some of the materials contained in “First hour/day with OCaml” but also significant changes. In contrast to the previous of this series, this tutorial is centered on the “working with files“ story: writing .ml and .mli files, modules, packages, compilation and dune.

However, as in the previous document, this a tour, not a detailed guide. This is not sufficient to be a proficient OCaml developer, but it is absolutely required to have an understanding of the materials presented here to start building a learning path into OCaml.