Adress comment on effects & unit by @christinerose
Initial text:
This function does not need to be passed any data to return its result. It is entirely context-dependent, entirely determined by external factors. Therefore, it could be passed anything, which would simply be discarded. Note it has to be passed something to trigger the computation, which is requesting the information to the operating system, as suggested by the module name. Since that function needs some data to trigger the computation but the data is going to be ignored, it makes sense to provide `()`, the unit value, the inhabitant of the type which only has a single value.

Proposed text:
No data needs to be passed to this function in order to return its result. It is entirely context-dependent, i.e., entirely determined by external factors. Therefore,I f it wasI passed anything, it would simply be discarded. Note that it has to be passed something to trigger the computation, since it requests the information from the operating system, as suggested by the module name. Since that function needs some data to trigger the computation but the data is going to be ignored, it makes sense to provide `()`, the unit value and inhabitant of the type that only has a single value.

This is taking some from the proposed, with extra edits.