Update about.eml
Really love this page overall. 

I thought it is currently underselling OCaml in significant ways. 
This PR changes a few minor things:

I rephrased the type inference part, since it did sound before like only the variables would be inferred. 

One thing that I removed, is the mention of the Java compiler. 
That one has received no commit for over 1 1/2 years, and it is unlikely, that it is picked up by anybody, so I changed it to the actually in active development happening ocaml-wasm project.

I also changed a few other things, and like to discuss some potential changes to this site here in the comments. 

Specifically, I think the GC is hugely undersold, since we use two different ones, who are both exceeding the expectations of people, who might be used to garbage collection in other languages. I would like to point out the dual nature of our GC, and how efficient it is. 

Real World OCaml does a good job with that.