Allow multiple modules in 1st program example lib
This discussion: https://discuss.ocaml.org/t/using-modules-15-characters-long-now has shown that the example project in "Your First OCaml Program" suffers from having a file called `lib/hello.ml`. Since it has the same name as the library defined in `lib/dune` (i.e. hello), it turns other file added in the `lib` folder to be private modules, which is confusing for (clever) newbies.

This patch attemps to resolve this issue the following way
- The file is now called `lib/world.ml` and `lib/world.mli`
- Constant exposed in `World` module is called `v`
- Gives rise to modules `Hello` and `Hello.World`

Drawback of this solution is exposure of Dune's wrapping mechanism.