Cookbook recipte for matrix add and mul
Implementation of the matrix addition and multiplication using the
Array module of the standard library and integer arrays. Note that
it relies on vector addition and dot product to define the matrix
operations in a functional style. It can be useful for those cases
where Owl library can't be used, otherwise these operations of the
matrices are defined in a, probably, more performant way in Owl.
Easy changes to implement it for floats.