(docs) Add opam env instructions for nushell to the windows page (#3307)
The powershell syntax is pretty close to what nushell wants. It'd be cool if `opam env` natively supported nushell but this is not too bad in the mean time.

I added this to the Windows page since I think it's one of the only shells Windows users are likely to be using other than cmd, Powershell, and Bash (which is trivial).